Leave clipboard stuff in user-facing module
The clipboard stuff can live entirely in user space, there's no need to
have its calls cross the userspace <-> Ammonite internals barrier.

Later on, this is helpful to Java-ize ReplAPI.
