+++ amm/repl/api/src/main/scala/ammonite/repl/api/ReplAPI.scala
@@ -214,17 +214,3 @@ trait Session{
     */
   def delete(name: String): Unit
 }
-trait Clipboard{
-  /**
-    * Reads contents from the system clipboard.
-    * @return System clipboard contents if they are readable as `String`,
-    *         empty string otherwise.
-    */
-  def read: String
-  /**
-    * Sets the contents of the system clipboard.
-    *
-    * @param data New contents for the clipboard.
-    */
-  def write(data: Internals.Writable): Unit
-}
