+++ ops/src/main/scala/ammonite/ops/FileOps.scala
@@ -29,27 +29,6 @@ object Internals{
       }
     }
   }
-
-
-  class Writable(val writeableData: geny.Generator[Array[Byte]])
-
-  object Writable extends LowPri{
-    implicit def WritableString(s: String) = new Writable(
-      geny.Generator(s.getBytes(java.nio.charset.StandardCharsets.UTF_8))
-    )
-    implicit def WritableBytes(a: Array[Byte]): Writable = new Writable(geny.Generator(a))
-
-  }
-  trait LowPri{
-
-    implicit def WritableGenerator[M[_], T](a: M[T])
-                                           (implicit f: T => Writable,
-                                            i: M[T] => geny.Generator[T]) = {
-      new Writable(
-        i(a).flatMap(f(_).writeableData)
-      )
-    }
-  }
 }
 
 /**
